Use HEAD to set the parent of the new commit.
In the first commit, the `Commit` class did not record the ID of its
parent commit. This was unnecessary since there were no previous
commits. However, once `.git/HEAD` exists, each newly created commit
should read it and store its contents as the ID of its parent.

Here we introduce a very basic implementation of this feature, and a new
abstraction for managing references: the `Refs` class. This has two
methods: one for reading the current HEAD value (returning `nil` if
there is no HEAD) and one for updating the value of HEAD.