Safely update HEAD by using a lockfile
Writing directly to `.git/HEAD` to update the current commit ID is
unsafe. Since writes are not atomic, a process may read the file while
it is being overwritten and get an empty string or a partial ID.

In the database, we fixed this by creating a file called
`tmp_obj_XXXXXX`, writing the object's data to that file, and then
moving it to its correct location. However, we must be more careful when
changing `.git/HEAD` because, unlike database files whose names are
determined by their content and so never change, this file has its
content repeatedly updated, and two concurrent processes might attempt
to write different data to it, creating a race condition.

Therefore, if one process tries to change `.git/HEAD` while another
process is also doing so, this probably indicates an error. Here we
introduce a mechanism for allowing files to be locked for updates, by
writing to a known path (the filename with `.lock` appended) that each
process must open with the `CREAT` and `EXCL` flags. If this fails, that
means some other process is currently changing the file.

This mechanism is encapsulated in the `Lockfile` class and will
eventually be used to implement other features, such as locking the
index and configuration files while they are being updated.