Don't overwrite database objects
Now that we're forming a chain of commits, some of the blobs will be the
same from one tree to the next. It's wasteful to spend time writing
files to `.git/objects` when they already exist, so we'll put in a check
for the object existing before beginning to write it.

This is also part of the mitigation against SHA-1 collisions being
exploited; we always keep the object we have rather than overwriting it
with something else.