Record the mode of executable files
So far, all tree entries have had their mode hard-coded to `100644`, the
mode of regular files. In order to store the fact a file is executable,
we must store it using the mode `100755`.

Here we achieve that by calling `stat(2)` on each file when we store
blobs into the database, and passing the stat into the `Entry` class.
When we serialise a `Tree` object, rather than hard-coding the mode of
each entry as `100644`, we ask the entry for its mode, which it chooses
by using the file stat to check if the file is executable by the current
user.