An `add` command for writing a single file to the index.
Here we begin to build up the machinery necessary for the `add` command.
In this first pass, it takes a single file path, writes that file's
contents to the database, and stores the resulting ID and stat data in
the index.

The process overwrites whatever the index currently contains. We don't
yet have code for reading the index and modifying it; we just build a
new empty `Index` in memory, add a single entry to it, and write it to
disk.

However, the serialisation logic present here is largely complete: the
`Index::Entry` class writes all the necessary stat information out,
including the file mode for executables, the blob ID is recorded
correctly, and the index is finished with a SHA-1 hash. Comparison of
this implementation with Git shows that the resulting `.git/index`
contents are byte-for-byte identical after adding a single file.

The index is not yet used to build commits -- the `commit` command still
reads directly from the working tree.