Allow `add` to take multiple filenames.
We'd like to be able to add all the files in the project to the index.
Since we don't yet support incremental updates to the index, one easy
way to support this would be to allow the `add` command to take multiple
filenames. We can then run a command like:

    $ ./bin/jit add $(find bin lib -type f)

to write an index containing all the project files.

A small change to the `Index` class is necessary once we have multiple
entries: the entries must be kept sorted by their name. We can achieve
this by storing the names in a `SortedSet` alongside the main entries
table.