Incrementally update the index
Rather than overwriting whatever is currently stored in `.git/index`,
the `add` command should read whatever is on disk, modify it, and then
write it back, all while holding a lock on the index file.

Here we add parsing logic to read the index data we've been writing
using the last few commits, taking care to validate the signature,
version and checksum as we go. We're currently ignoring the path length
field and just scanning in blocks of 8 bytes until we find the null
terminator.