Use the index to build the next commit
Here we migrate the `commit` command from using the `Workspace` to using
the `Index` to build the tree for the new commit. Rather than reading
the entire project from disk and hashing it, we read the cached blob IDs
from `.git/index` and use those to build a tree.

The data structures created by `Index` and `Index::Entry` are quite
similar to the original `Entry` class. The only difference is the
addition of some convenience methods for accessing the basename and
parent directories of the entry's path.