Begin adding unit tests
To get a simple test suite in place, this adds a single test for adding
an entry to the index. The tests are written using Minitest and run
using Rake, both parts of the Ruby standard library.

We don't need to clean up files after the test because we don't tell the
index to write to disk at any point.