Discard index files when replaced with directories
If the index contains the following files:

    alice.txt
    bob.txt

and a file named `alice.txt/nested.txt` is added, then the entry
`alice.txt` should be removed, since that name now corresponds to a
directory containing the new file.

We accomplish this by iterating over the newly added entry's parent
directories, and removing any entries whose names are equal to any of
them.