Handle file-related errors in the `add` command
When the user passes arguments to the `add` command, those inputs might
refer to filenames that do not exist, or that point to files that are
not readable. This set of changes makes sure such cases are handled
gracefully.

First, the `Workspace` class is modified to raise errors from its
`list_files` method if the given path does not name an existing file,
and from `read_file` and `stat_file` if the given path is not readable.
These errors are then caught in the `add` command code and handled so
that nothing is added to the database if any of the inputs name a
missing file.

Premature exit from `add` could leave the file `.git/index.lock` in
place before anything is written to it, preventing further calls from
acquiring the lock. Therefore the error handlers make sure to remove
this file before the program exits.