Handle a stale `index.lock` existing when `add` is called
If `.git/index.lock` already exists when we want to load the index, an
error should be raised. Here we change `Lockfile` so that
`hold_for_update` raises an exception rather than returning `false` if
the lock cannot be acquired.

This exception is then caught by the `add` command code, which prints
instructions for the user to remove the lock if desired.