Extract command code into classes
The code in `bin/jit` would benefit from being encapsulated in classes
so we can execute it from a test suite. This commit makes the first step
of this process by transplanting the `bin/jit` code verbatim from its
current location into one class for each command: `Command::Init`,
`Command::Add` and `Command::Commit`.

The command classes contain exactly the same code as was previously in
`bin/jit` and will require further changes to make it easier to test.
Those changes will be done in separate commits so they can be
distinguished from this change that just moves a lot of code around.