Replace global references in `Command` classes with injected dependencies
To decouple the `Command` classes from the global environment and make
them easier to test, we replace all references to global state and
objects with instance variables that are passed into the command class
by the caller, which is `bin/jit` at runtime.

The replaced globals are:

- `Dir.getwd`, the process's current working directory
- `ENV`, the set of environment variables
- `ARGV`, the command-line arguments
- `$stdin`, the standard input stream
- `$stdout` and `$stderr`, the standard output and error streams

The `Command` classes now inherit from `Command::Base`, which stores
references to all the above. It also provides some convenience methods
for common tasks, as well as overriding some global methods so they work
correctly inside the commands. These are:

- `puts` writes to the command's stored `@stdout` stream, not the Ruby
  process's global stdout
- `exit` throws to stop the command's execution, but just stores the
  exit status rather than actually stopping the Ruby process