Refactor the `Command::Add` class
While adding error handling to the `add` command, we made it much more
complicated and hard to follow. Now that the code is in a class, and
covered by tests, we can refactor it into a better shape.

The error-handling code has been split into separate methods so it
doesn't clutter up the main body of the command. Then, the details of
each step of the command -- expanding the input paths, and adding each
to the index -- are also split into their own methods. The `run` method
is now just a high-level description of the steps taken by the command.