Begin the `status` command, listing untracked files
Here we start work on the `status` command. The ultimate aim is to list
out the changes to the index and the workspace, and the untracked files,
in both the porcelain and long formats supported by Git.

We start small with a single test that adds one untracked file to the
repository and checks that it's listed with the right status. In the
absence of other files, this can be made to work by just listing all the
files in the workspace with `??` next to them.

To implement all the underlying business logic we'll be testing using
the porcelain format, so for the time being that's all that the command
can generate.