Exclude tracked files from the untracked files list
The previous commit listed untracked files by simply listing all the
files in the workspace. Here we add another test with some files in the
index, to make sure they're not displayed as untracked by the `status`
command.

This is accomplished by loading the index (with `load` because we won't
be updating it), and filtering any files out of the workspace listing
that appear in the index.