Display untracked directories without their contents
This is an edge case in the display of untracked files. If a directory
does not contain any files that are in the index, then it itself should
be listed as untracked, not the files inside it.

Instead of asking `Workspace` for a full listing of all the files, we
instead now explore the workspace one directory at a time, using a new
method that returns the contents of one directory with the `stat()`
result for each file. Any items in the directory that are not tracked
are immediately added to the untracked list and not explored further.

This piggy-backs on the data structures added inside `Index` to deal
with file/directory name conflicts. We can easily tell if a directory is
indexed because its name will appear in these internal structures.