Don't list untracked directories that contain no trackable files
A further refinement of the previous edge case: if a directory is
untracked, it should only be listed if it contains any files that would
be added to the index if we passed the directory to the `add` command.
In the absence of features for ignoring files, this just means the
directory must, directly or indirectly, contain at least one file.

To handle this we use a new recursive procedure that, rather than trying
to discover the whole contents of a directory, merely tries to find a
single file in the directory and returns as soon as it does so. This
adds a little work to confirming a directory should be listed, but still
allows us to skip fully exploring it.