Detect workspace differences by size
Having completed listing of untracked files, we move on to listing
differences between the index and the workspace. Since the index caches
the `stat()` information of the file when it was last added, we can
detect some kinds of changes by comparing this information with the stat
that we just got while searching for untracked files. For example, if
the cached file size is different from the one we just read, then we
know the file has changed and we don't need to spend time reading it.

As this is the first kind of workspace change we're detecting, it comes
with a bunch of new structures for reporting them. Workspace-changed
files are added to a new set, and all are currently listed with the
status `M`.