Detect workspace differences by mode
This is an extension to the previous change that uses the index's cached
stat information to compare against the stat information gathered while
scanning the workspace.

In this case, if the file's mode has changed, then it should be listed
as modified. We don't compare the literal mode, since Git only stores
100644 and 100755; instead we calculate the mode the file would be given
if added to the index and compare that to the stored value.