Detect size-preserving workspace changes
While comparing the index to the workspace stat values, if the file's
size and mode have not changed, then we need to read its contents, hash
them to get an object ID, and compare that to the one stored in the
index. This is done by creating a `Blob` out of the file's current
contents, and asking the database to hash this object without saving it
to disk.

We have extracted the functionality to hash an object without saving it
in the `Database` interface. We'd like to avoid saving the object partly
because it avoids the work of compressing it and writing it to disk, but
also because any blobs created by this process are likely not to end up
in any final tree that ends up in a commit. With no references to these
blobs in the commit graph or index, they'll just get lost so there's no
point keeping them around.