Skip comparing a file's contents if its timestamps are unchanged
In the previous commit, we arranged things to that any files whose size
and mode are unchanged from the index are read and hashed and compared
to the indexed blob ID for the file. This is really time-consuming and
will usually result in most of the files in a project being read and
hashed just to decide they've not changed.

To fix this, we use the ctime and mtime cached in the index. If the
file's ctime and mtime are equal to the cached values, it almost
certainly hasn't changed and we can skip reading it. If we do read it
and it turns out not to have changed, then we update the stat
information in the index to avoid reading the file next time.

Since we're now updating the contents of the index, we need to read it
with `load_for_update` and write any updates back at the end of the
data-gathering phase.