Detect files deleted in the workspace
If a file that's listed in the index is no longer on disk, we should
list it as deleted with a `D` status. Since files can now have multiple
statuses, we store the state of each file as a `Set` and adjust our
existing change-recording code to note the type of change in each case:
`workspace_modified` or `workspace_deleted`.