Read commits, trees and blobs from the database
In order to support the next phase of developing the `status` command,
we need to be able to read the HEAD commit and get the blob IDs of
everything in its tree.

Here we implement routines for reading database objects, determining
their type, and then handing the rest of the file data off to the
relevant class for further parsing using a `StringScanner`.

The `Commit` class currently only exposes the `tree` property since
that's all we're going to need to read during the `status` command. It
doesn't bother to parse the `author` line into an `Author` object,
because we currently have no need to inspect that information in a
structured way. It also ignores the `committer` line, because when
writing commits we always make the author and committer the same. To
handle arbitrary real-world commits we will need to add support for
handling the committer field separately.

The `Tree` objects read from the database are unlike the ones used to
write to it; the latter are nested structures of `Index::Entry` values
while the former is a flat structure of the newly introduced
`Database::Entry` class. The new `Entry` class exists to model just the
fields stored in trees, which lack a lot of the information stored in
the index, but its interface is similar enough to `Index::Entry` in the
ways that `Tree` cares about. We don't read trees as nested structures
because many operations won't require us to read an entire tree; we'd
like the caller to control the recursive search just as the `status`
code does when looking for untracked files.

I have decided to implement support for loading `Blob` objects here even
though it's not needed for the `status` command, because it's the
simplest database object type to read.