Refactor the `Command::Status` state to support the long format
The way data is currently organised in this command does not make it
particularly easy to implement the long status format, where index and
workspace changes are listed separately. Here's an example of the
current data organisation:

    @changes = {
      "1.txt"     => Set.new([:index_modified, :workspace_deleted]),
      "a/2.txt"   => Set.new([:index_deleted]),
      "a/b/3.txt" => Set.new([:workspace_modified]),
      "z.txt"     => Set.new([:index_added])
    }

This structure requires much iteration through it to discover which
items should be listed as index changes and with what status, and
likewise for the workspace. A better organisation that still keeps the
porcelain format easy to write, is:

    @index_changes = {
      "1.txt"   => :modified,
      "a/2.txt" => :deleted,
      "z.txt"   => :added
    }

    @workspace_changes = {
      "1.txt"     => :deleted,
      "a/b/3.txt" => :modified
    }

This commit refactors the class to this design so that adding the long
format will be more straightforward. In the process the code for the
existing porcelain format is simplified.