Implement the long status format
The default output of the `status` command should not be the porcelain
format, but the long one, listing index changes, workspace changes and
untracked files in three distinct sections.

Following from the previous refactoring, here we implement that format
by listing out the contents of the various data structures generated by
the scan of the repository, using a different format to the one already
implemented.

This format omits some of the helpful instructions that Git normally
includes because we don't yet have support for all the requisite
commands.

The porcelain format is no longer the default; the argument
`--porcelain` must be passed to get this instead.