Display index changes in the correct order
This fixes an edge case in the long status format. Because Ruby hashes
maintain insertion order and files deleted between HEAD and the index
are detected last, such files are always listed last in the long status
format. Added and modified files are listed in the right order, because
we detect them in the order they appear in the index.

To keep everything in order, we introduce a new subclass of `Hash` that
uses a `SortedSet` to keep its keys sorted. By using this to store index
and workspace changes, we make sure everything is printed in the right
order.