The Myers diff algorithm
Here we implement the main algorithm described in "An O(ND) Difference
Algorithm and its Variations" [1] for finding a shortest edit script for
converting one string into another.

The particular representation of strings used here is that of lines in a
text file; each input is split into its constituent lines before being
fed into the algorithm. If the inputs are already arrays, they are left
unchanged.

Calling `Diff.diff(a, b)` returns an array of `Diff::Edit` objects
describing the changes between the two versions, which can be
immediately printed to produce a diff representation in the terminal.

[1]: http://www.xmailserver.org/diff2.pdf