Display diff headers for workspace-modified files
The previous two commits put some infrastructure in place for building
the `diff` command: the Myers diff algorithm, and a reusable version of
the `status` command's business logic. We can now start building the
`diff` command.

We start with displaying header information about one type of change:
files that are modified between the index and workspace. We display the
file's path and its object ID in the index and workspace.

Object IDs are shortened to 7 digits using a new method in `Database`.
In future it would be good for this method to ensure that the truncated
object ID still identifies a single object and extend it otherwise.