Display diff headers for files with changed modes
This is an edge case to the functionality introduced for displaying diff
headers. If a file's mode is changed, the old and new mode are displayed
on their own lines rather than as part of the `index` line.

To support this, `Repository::Status` now exposes its `stats` variable
that stores the `File::Stat` for every file we already read from the
workspace.