Display diff headers for files deleted from the workspace
When a file has been removed from the workspace but is still in the
index, `diff` should print it with various items set to null-values in
the diff headers. In particular, its object ID is `0000000`, and its
path for the diff is `/dev/null`.

The method for generating this output is simpler than that for modified
files, because the mode of deleted files will always differ from that
in the index, so some conditions the are present in the modified-file
path are not needed here.