Extract creation of `Diff::Target` objects
In the previous commit we extracted the display logic from
`diff_file_{modified,deleted}` and turned them into functions that each
construct two `Target` objects and pass them to `print_diff`. This has
exposed another similarity, which is that the patterns for constructing
`Target` objects are common, especially since both methods construct a
target from the index.

Here we extract three methods for building `Target` objects:
`from_index`, `from_workspace` and `from_nothing`, which build all the
types of targets currently in use. This hollows out the
`diff_file_{modified,deleted}` methods to the extent that their effect
can be inlined into the main loop of the command.