Display diff headers for index-modified files
Having built out the mechanics of the `diff` command in previous
commits, we now extend it with the `--cached` switch to make it display
differences between the HEAD commit and the index. Just as for
index/workspace differences, we iterate over the changes returned by
`Repository::Status`, building pairs of `Target` objects and sending
them to `print_diff`.

The only state handled so far is files that are modified between HEAD
and the index. Other states are handled in subsequent commits.

This has necessitated two additions: the `from_head` method builds
`Target` objects from files in the HEAD commit's tree, and this relies
on `Repository::Status` exposing its `head_tree` variable so that we can
reuse objects read from the database while computing the status.