Add line number information to `Diff::Edit` objects
In order to make it easier to calculate and display hunk boundaries, it
will be useful to have the output of `Diff.diff` contain line number
information so we don't need to reconstruct it by walking through the
edits.

To that end, we replace the existing `text` field of `Diff::Edit` with
two fields: `a_line` and `b_line`, each of which point to `Diff::Line`
objects that have a `number` and `text`. This makes it much easier to
tell which line numbers in the output match, since edits with type
`:eql` have both an `a_line` and a `b_line`.

When `Diff.diff` calls `Diff.lines`, this helper method now wraps each
line of the file in a `Line` object with a 1-indexed number.

Some minor changes to `Diff::Myers` are needed to cope with these new
structures, but nothing in `Command::Diff` needs to change since it only
relies on the `Diff::Edit#to_s` method.