Split diffs into hunks
Currently, the `diff` command prints the entirety of each changed file.
Here we make it filter out long sections of unchanged code and display
changes in "hunks", which are regions of a file containing nearby
changes. Edits are displayed with three lines of surrounding context,
and two edits are considered "nearby" if their context overlaps.

Each hunk is printed with a header that says where it begins in each
version of the file, and how many lines from each version it contains.