Parse author information from commits
When a revision expression includes a short object ID, and that prefix
matches multiple objects, Git will print an error stating that the
requested branch start point is ambiguous. For example:

    error: short SHA1 bc24 is ambiguous
    hint: The candidates are:
    hint:   bc247fb commit 2018-01-04 - An `add` command for writing a single file to the index
    hint:   bc24976 tree
    fatal: Not a valid object name: 'bc24'.

In order to display this information, we need to be able to pluck the
author date and the first line of the message from a commit.

Until now, the `Database::Commit` class did not expose the `author`
attribute, so there was no need to parse that header. Now, instead of
leaving it as a string, we parse it into an `Author` object and expose
it.

To facilitate printing the above message, two new methods are added to
the database objects: `Author#short_date` and `Commit#title_line`.