Print errors for revisions that resolve to non-commits
It's possible to supply a revision argument to the `branch` command that
identifies some object that's not a commit, for example the ID of a tree
or blob. In this situation, the `branch` command should report that the
object is not usable as a start point.

We do this by passing an argument to `Revision#resolve` indicating which
type of object we expect. If the resolved object is not of the requested
type, an error is raised.